Fix code to be compliant with new standard
All Google extensions going forward have to comply with specific
development starts to protect from XSS and other attacks. This required
that all the inline JS be moved out to a standalone file, and for all
inline event handlers to be replaced with eventListeners.

The link to the new Content Security Poliy is here:
https://developer.chrome.com/extensions/contentSecurityPolicy